refactor: improve CLI output formatting
Truncate displayed lines to the first 100 characters using firstchars. Remove the unused checkError function.
