+++ main.go
@@ -8,6 +8,7 @@ import (
 	"github.com/ebisuG/search-all-user-bookmark/internal/config"
 	"github.com/ebisuG/search-all-user-bookmark/internal/infra"
 	"github.com/ebisuG/search-all-user-bookmark/internal/search"
+	"github.com/mattn/go-runewidth"
 	"github.com/muesli/termenv"
 
 	"github.com/charmbracelet/bubbles/textinput"
@@ -152,9 +153,33 @@ func (m model) View() string {
 
 func (r result) FormatDisplay() {
 	nameColor := lipgloss.Color("#F77F0F")
+	lineLength := 100
 
 	for _, v := range r {
-		hypelink := termenv.Hyperlink(v.bookmarkUrl.record.raw, v.bookmarkTitle.record.raw)
+		hypelink := termenv.Hyperlink(v.bookmarkUrl.record.raw, firstChars(lineLength, v.bookmarkTitle.record.raw))
 		fmt.Println(" ", lipgloss.NewStyle().Foreground(nameColor).Bold(false).Render(hypelink))
 	}
+}
+
+func firstChars(fixedLength int, s string) string {
+	result := ""
+	currentWidth := 0
+	ellipsis := "..."
+
+	for _, r := range s {
+		w := runewidth.RuneWidth(r)
+		if currentWidth+w >= fixedLength-len(ellipsis) {
+			result += ellipsis
+			currentWidth += len(ellipsis)
+			break
+		}
+		result += string(r)
+		currentWidth += w
+	}
+
+	if currentWidth <= fixedLength {
+		result += runewidth.FillRight("", fixedLength-currentWidth)
+	}
+
+	return result
 }
